current weather -> successfully fetched current weather conditions

+++ app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Coord.java
@@ -6,23 +6,23 @@ import com.google.gson.annotations.SerializedName;
 public class Coord {
 
     @SerializedName("lat")
-    private Long mLat;
+    private Double mLat;
     @SerializedName("lon")
-    private Long mLon;
+    private Double mLon;
 
-    public Long getLat() {
+    public Double getLat() {
         return mLat;
     }
 
-    public void setLat(Long lat) {
+    public void setLat(Double lat) {
         mLat = lat;
     }
 
-    public Long getLon() {
+    public Double getLon() {
         return mLon;
     }
 
-    public void setLon(Long lon) {
+    public void setLon(Double lon) {
         mLon = lon;
     }
